requirements-dev.txt
# Development dependencies for Lambda Coverage Layer

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
moto>=4.2.0  # For mocking AWS services

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# CDK dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Build and packaging
wheel>=0.40.0
setuptools>=65.0.0